[Rename] 자원서버 프로젝트명 수정
