+++ resource-server/settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'resource-server'
+rootProject.name = 'gold-resource'
