+++ resource-server/src/main/java/com/ryuneng/goldresource/GoldResourceApplication.java
@@ -1,14 +1,14 @@
-package wantedgold.resourceserver;
+package com.ryuneng.goldresource;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class ResourceServerApplication {
+public class GoldResourceApplication {
 
     public static void main(String[] args) {
 
-        SpringApplication.run(ResourceServerApplication.class, args);
+        SpringApplication.run(GoldResourceApplication.class, args);
     }
 
 }
