+++ resource-server/src/main/java/com/ryuneng/goldresource/domain/order/repository/OrderRepository.java
@@ -1,7 +1,7 @@
-package wantedgold.resourceserver.domain.order.repository;
+package com.ryuneng.goldresource.domain.order.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import wantedgold.resourceserver.domain.order.entity.Order;
+import com.ryuneng.goldresource.domain.order.entity.Order;
 
 public interface OrderRepository extends JpaRepository<Order, Long> {
