+++ resource-server/src/main/java/com/ryuneng/goldresource/domain/product/repository/ProductRepository.java
@@ -1,7 +1,7 @@
-package wantedgold.resourceserver.domain.product.repository;
+package com.ryuneng.goldresource.domain.product.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import wantedgold.resourceserver.domain.product.entity.Product;
+import com.ryuneng.goldresource.domain.product.entity.Product;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
