+++ resource-server/src/main/java/com/ryuneng/goldresource/global/config/JpaAuditingConfig.java
@@ -1,4 +1,4 @@
-package wantedgold.resourceserver.global.config;
+package com.ryuneng.goldresource.global.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
